[Support] Add `sphinx` dependency and create "docs" group.
This commit updates the dependencies in the pyproject.toml file by adding the `sphinx` package to the [tool.poetry.group.docs.dependencies] section. Additionally, it modifies the Makefile to include the `--with docs` flag when running the `poetry install --all-extras` command. These changes ensure that the documentation dependencies are properly installed and up to date.